Linux 32bit build fixed.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@3817 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/core/PWSAuxParse.cpp
@@ -345,7 +345,7 @@ StringX PWSAuxParse::GetAutoTypeString(const StringX &sx_in_autotype,
             // Send the lot
             sxtmp += sx_notes;
           } else
-          if (line_number <= static_cast<int>(vsxnotes_lines.size())) {
+          if (line_number <= vsxnotes_lines.size()) {
             // User specifies a too big a line number - ignore the lot
             sxtmp += vsxnotes_lines[line_number - 1];
           }
